+ blogspot (RSS feed) scrapers, - deprecated scrapers
