+++ src/Scrapers/TopProxysScraper.php
@@ -1,0 +1,19 @@
+<?php declare(strict_types = 1);
+
+namespace Vantoozz\ProxyScraper\Scrapers;
+
+/**
+ * Class TopProxysScraper
+ * @package Vantoozz\ProxyScraper\Scrapers
+ */
+final class TopProxysScraper extends AbstractRssBloggerScraper
+{
+
+    /**
+     * @return string
+     */
+    protected function rssBloggerUrl(): string
+    {
+        return 'https://topproxys.blogspot.com/feeds/posts/default';
+    }
+}
