--- tests/unit/Scrapers/SpysMeScraperTest.php
@@ -1,92 +1,0 @@
-<?php declare(strict_types = 1);
-
-namespace Vantoozz\ProxyScraper\UnitTests\Scrapers;
-
-use PHPUnit\Framework\TestCase;
-use Vantoozz\ProxyScraper\Enums\Metrics;
-use Vantoozz\ProxyScraper\Exceptions\HttpClientException;
-use Vantoozz\ProxyScraper\HttpClient\HttpClientInterface;
-use Vantoozz\ProxyScraper\Proxy;
-use Vantoozz\ProxyScraper\Scrapers\SpysMeScraper;
-
-/**
- * Class SpysMeScraperTest
- * @package Vantoozz\ProxyScraper\Scrapers
- */
-final class SpysMeScraperTest extends TestCase
-{
-    /**
-     * @test
-     * @expectedException \Vantoozz\ProxyScraper\Exceptions\ScraperException
-     * @expectedExceptionMessage error message
-     */
-    public function it_throws_an_exception_on_http_client_error(): void
-    {
-        /** @var HttpClientInterface|\PHPUnit_Framework_MockObject_MockObject $httpClient */
-        $httpClient = $this->createMock(HttpClientInterface::class);
-        $httpClient
-            ->expects(static::once())
-            ->method('get')
-            ->willThrowException(new HttpClientException('error message'));
-
-        $scraper = new SpysMeScraper($httpClient);
-        $scraper->get()->current();
-    }
-
-    /**
-     * @test
-     */
-    public function it_returns_source_metric(): void
-    {
-        /** @var HttpClientInterface|\PHPUnit_Framework_MockObject_MockObject $httpClient */
-        $httpClient = $this->createMock(HttpClientInterface::class);
-        $httpClient
-            ->expects(static::once())
-            ->method('get')
-            ->willReturn("222.111.222.111:8118\n111.222.111.222:8118");
-
-        $scraper = new SpysMeScraper($httpClient);
-        $proxy = $scraper->get()->current();
-
-        static::assertInstanceOf(Proxy::class, $proxy);
-        /** @var Proxy $proxy */
-        static::assertSame(Metrics::SOURCE, $proxy->getMetrics()[0]->getName());
-        static::assertSame(SpysMeScraper::class, $proxy->getMetrics()[0]->getValue());
-    }
-
-    /**
-     * @test
-     */
-    public function it_returns_a_proxy(): void
-    {
-        /** @var HttpClientInterface|\PHPUnit_Framework_MockObject_MockObject $httpClient */
-        $httpClient = $this->createMock(HttpClientInterface::class);
-        $httpClient
-            ->expects(static::once())
-            ->method('get')
-            ->willReturn("222.111.222.111:8118\n111.222.111.222:8118");
-
-        $scraper = new SpysMeScraper($httpClient);
-        $proxy = $scraper->get()->current();
-
-        static::assertInstanceOf(Proxy::class, $proxy);
-        static::assertSame('222.111.222.111:8118', (string)$proxy);
-    }
-
-    /**
-     * @test
-     */
-    public function it_skips_bad_rows(): void
-    {
-        /** @var HttpClientInterface|\PHPUnit_Framework_MockObject_MockObject $httpClient */
-        $httpClient = $this->createMock(HttpClientInterface::class);
-        $httpClient
-            ->expects(static::once())
-            ->method('get')
-            ->willReturn('2312318');
-
-        $scraper = new SpysMeScraper($httpClient);
-
-        static::assertNull($scraper->get()->current());
-    }
-}
